+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `View.track` to execute (async) actions that will trigger the invalidation of the `View`.
 - `View.escape` to escape the tracker zone for `EventHandler`s that are registered inside the `View`.
 - Expose `View` in `BuildContext`.
+- Experimental `SubView`.
 
 ## 0.2.0+1